Update instruction view
